add/x11/VBoxClient: Fix warning. bugref:10348
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@98202 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/x11/VBoxClient/display-svga-xf86cvt.cpp
@@ -197,7 +197,7 @@ DisplayModeR VBoxClient_xf86CVTMode(int HDisplay, int VDisplay, float VRefresh /
         (void) VBackPorch;
 
         /* 11. Find total number of lines in vertical field */
-        Mode.VTotal = VDisplayRnd + 2 * VMargin + VSyncAndBackPorch + Interlace + CVT_MIN_V_PORCH;
+        Mode.VTotal = (int)(VDisplayRnd + 2 * VMargin + VSyncAndBackPorch + Interlace + CVT_MIN_V_PORCH);
 
         /* 5) Definition of Horizontal blanking time limitation */
         /* Gradient (%/kHz) - default 600 */
